pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["Sameer Shahi Thakuri (Sam) <70689865+Sam231221@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
